add Days type and use for Medicare levy surcharge
Add the Days type for recording the number of days in a year.
Update TaxReturnInfo and related types and functions to use it.  In
particular, the tax assessment calculation now implements part-year
Medicare levy surcharge exemption.

Remove the HasMLSExemption, HasHelpBalance and HasSfssBalance
classes.  These were supposed to cover PAYG withholding declarations
as well as tax returns, but the types turned out to be different so
the classes are going away.