apply wholeDollars to income components
As far as I can tell, the various income components all get
truncated.  Deductions get subtracted, then the result of that also
gets truncated to give the final value for taxable income.

Maybe I am wrong, but this method gives me the same value as the ATO
assessment, so I'll go with it for now.